Model/requirements.txt
torch>=2.1.2
torchaudio>=2.1.2
torchvision>=0.16.2
numpy>=1.24.4
scikit-learn>=1.3.2
pandas>=2.0.3
librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1
opensmile>=2.4.2
transformers>=4.36.2
sentence-transformers>=2.2.2
fastapi>=0.104.1
uvicorn>=0.24.0
slowapi>=0.1.9
python-dotenv>=1.0.0
pyyaml>=6.0.1
sqlalchemy>=2.0.23
tqdm>=4.66.1
joblib>=1.3.2
pydantic>=2.5.3
requests>=2.31.0
python-multipart>=0.0.6
h5py>=3.10.0
matplotlib>=3.8.2